First commit initialization

+++ src/com/javaguru/shoppinglist/service/ProductService.java
@@ -1,32 +1,47 @@
 package com.javaguru.shoppinglist.service;
 
 import com.javaguru.shoppinglist.domain.Product;
+import com.javaguru.shoppinglist.exeptions.ItemNotFoundException;
+import com.javaguru.shoppinglist.exeptions.UserValidationException;
 import com.javaguru.shoppinglist.repository.ProductRepositoryImpl;
-import com.javaguru.shoppinglist.service.ProductValidationService;
 
-public class ProductService  {
+public class ProductService {
 
     private ProductRepositoryImpl repository = new ProductRepositoryImpl();
-    private ProductValidationService validationService = new ProductValidationService();
-
+    private ProductValidationService productValidate = new ProductValidationService();
 
     public Long addProduct(Product product) {
+        productValidate.checkNotNull(product);
         Product createdProduct = repository.addProduct(product);
         return createdProduct.getId();
     }
 
     public Product findProduct(Long productId) {
-        return repository.findProduct(productId);
+        Product product = repository.findProduct(productId);
+        if (product == null) {
+            throw new ItemNotFoundException("Product with + " + productId + " ID not found");
+        }
+        return product;
 
     }
-    public boolean deleteProduct(Long idToDelete){
-        return repository.deleteProduct(idToDelete);
+
+    public Boolean deleteProduct(Long idToDelete) {
+        if (idToDelete == null) {
+            throw new UserValidationException("Wrong product id, its not possible to delete product");
+        }
+        Boolean product = repository.deleteProduct(idToDelete);
+        return product;
 
     }
 
-    public Product updateProduct(Long productId) {
-        return repository.updateProduct(productId);
+    public Product updateProduct(Long productId, Product product) {
+        if (product == null) {
+            throw new UserValidationException("Wrong product id, its not possible to update your product");
+        }
+        Product newProduct = repository.updateProduct(productId, product);
+        return newProduct;
     }
+
